project-health-check/requirements.txt
amqp==5.3.1
asgiref==3.11.1
async-timeout==5.0.1
billiard==4.2.4
celery==5.6.2
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==5.2.11
django-crontab==0.7.1
django-environ==0.13.0
django-filter==25.2
django-jazzmin==3.0.3
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
exceptiongroup==1.3.1
gunicorn==25.1.0
idna==3.11
kombu==5.6.2
mysqlclient==2.2.8
packaging==26.0
prompt_toolkit==3.0.52
PyJWT==2.11.0
python-dateutil==2.9.0.post0
redis==7.2.0
requests==2.32.5
ruff==0.15.2
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
whitenoise==6.6.0
